flask-backend/requirements.txt
# Core packages
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7

# Database and ORM
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5
Flask-Script==2.0.6
alembic==1.12.1

# Authentication
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Session==0.5.0
redis==5.0.1

# CORS and Requests
Flask-Cors==4.0.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
uuid==1.30
gunicorn==21.2.0

# Development
pytest==7.3.1
pytest-cov==4.1.0
black==23.3.0
flake8==6.0.0
isort==5.12.0